requirements.txt
apsw==3.46.1.0
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
mysql-connector-python==9.0.0
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
python-dotenv==1.0.1
setuptools==73.0.0
sqlparse==0.5.1
tzdata==2024.1
wheel==0.44.0
whitenoise==6.7.0
